Recent delete endpoints
