+++ Library-backend/package.json
@@ -2,12 +2,10 @@
   "name": "library",
   "version": "1.0.0",
   "description": "House of Pages backend Project",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon index.js",
-    "seed": "",
-    "start": ""
+    "start": "nodemon server.js",
+    "seed": "PGPASSWORD=goodworks17 psql -U postgres -d postgres -a -f ../Library-backend/library.sql"
   },
   "author": "Team 3",
   "license": "ISC",
